requirements_cpu.txt
langchain-ollama
streamlit
paddleocr
paddlepaddle
setuptools
python-dotenv

#groq
groq